Add unified custom cell hook to make things easier
