+++ packages/cells/src/index.ts
@@ -1,4 +1,4 @@
-import { GridCellKind, DataEditorProps, ProvideEditorCallback, GridCell } from "@glideapps/glide-data-grid";
+import {  DataEditorProps, ProvideEditorCallback, GridCell, useCustomCells } from "@glideapps/glide-data-grid";
 import StarCellRenderer from "./cells/star-cell";
 import SparklineCellRenderer from "./cells/sparkline-cell";
 import TagsCellRenderer from "./cells/tags-cell";
@@ -6,7 +6,6 @@ import UserProfileCellRenderer from "./cells/user-profile-cell";
 import DropdownCellRenderer from "./cells/dropdown-cell";
 import ArticleCellRenderer from "./cells/article-cell";
 import RangeCellRenderer from "./cells/range-cell";
-import * as React from "react";
 
 type DrawCallback = NonNullable<DataEditorProps["drawCell"]>;
 
@@ -24,28 +23,15 @@ export function useExtraCells(): {
     drawCell: DrawCallback;
     provideEditor: ProvideEditorCallback<GridCell>;
 } {
-    const drawCell = React.useCallback<DrawCallback>(args => {
-        const { cell } = args;
-        if (cell.kind !== GridCellKind.Custom) return false;
-        for (const c of cells) {
-            if (c.isMatch(cell)) {
-                return c.draw(args, cell as any);
-            }
-        }
-        return false;
-    }, []);
-
-    const provideEditor = React.useCallback<ProvideEditorCallback<GridCell>>(cell => {
-        if (cell.kind !== GridCellKind.Custom) return undefined;
-
-        for (const c of cells) {
-            if (c.isMatch(cell)) {
-                return c.provideEditor(cell as any) as ReturnType<ProvideEditorCallback<GridCell>>;
-            }
-        }
-
-        return undefined;
-    }, []);
-
-    return { drawCell, provideEditor };
+    return useCustomCells(cells);
 }
+
+export { 
+    StarCellRenderer as StarCell,
+    SparklineCellRenderer as SparklineCell,
+    TagsCellRenderer as TagsCell,
+    UserProfileCellRenderer as UserProfileCell,
+    DropdownCellRenderer as DropdownCell,
+    ArticleCellRenderer as ArticleCell,
+    RangeCellRenderer as RangeCell,
+ };
