+++ packages/core/src/index.ts
@@ -14,5 +14,6 @@ export { parseToRgba } from "./data-grid/color-parser";
 export { measureTextCached } from "./data-grid/data-grid-lib";
 export { getDataEditorTheme as getDefaultTheme } from "./common/styles";
 export { useColumnSizer } from "./data-editor/use-column-sizer";
+export { useCustomCells, CustomCellRenderer } from "./data-editor/use-custom-cells";
 
 export default DataEditor;
